Fix the validation of incompatible options
All clients must be validated, not only the first one (which was the case
due to returning during the first iteration of the loop over clients).
The easiest way to handle it is to put the validation inside the prototype
node, so that each client is validated separately. This also allows to have
a better error message (as the error message will tell you which client is
invalid as the error will be inside the client configuration node).
The error message is also fixed to talk about the third incompatible features.